// client/src/interface/settings.d.ts
export interface UserType {
    id: number,
    username: string,
    email: string,
    name: string,
    surname: string,
    phone: string,
    country: string,
    city: string,
    address: string,
    gender: string,
    date_of_birth: string,
    password: string,
}

export interface SeetingsFieldType {
    email: string
}